Safari-611.1.10/LayoutTests/fast/events/autoscroll-main-document-expected.txt

Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
This test ensures that drag-selecting works correctly in the main frame.

PASS firstRange.startOffset is 18
PASS firstRange.endOffset is 425
PASS successfullyParsed is true

TEST COMPLETE
